fixes #3: Service Discovery on External Components failed
The implementation queries for disco#info and disco#items, to base it's agentinformation responses on. The disco#info/item requests that were performed by the plugin were implemented using a shortcut that is only valid for _internal_ components.

This commit ensures that when discovering information of external components, a full XMPP request is performed.